# Small vector optimization for operands.