--- a/test/util/CMakeLists.txt
+++ b/test/util/CMakeLists.txt
@@ -20,3 +20,7 @@ add_spvtools_unittest(TARGET bit_vector
   SRCS bit_vector_test.cpp
   LIBS SPIRV-Tools-opt
 )
+
+add_spvtools_unittest(TARGET small_vector
+  SRCS small_vector_test.cpp
+)