… test/util/small_vector_test.cpp