We replace the std::vector in the Operand class by a new class that does
a small size optimization.  This helps improve compile time on Windows.

Tested on three sets of shaders.  Trying various values for the small
vector.  The optimal value for the operand class was 2.  However, for
the Instruction class, using an std::vector was optimal.  Size of "0"
means that an std::vector was used.


<table>
  <tr>
   <td>OperandData Size
   </td>
   <td colspan="3" >OperandList size
   </td>
  </tr>
  <tr>
   <td>
   </td>
   <td>0
   </td>
   <td>4
   </td>
   <td>8
   </td>
  </tr>
  <tr>
   <td>0
   </td>
   <td>489s (baseline)    
   </td>
   <td>544s
   </td>
   <td>684s
   </td>
  </tr>
  <tr>
   <td>1
   </td>
   <td>593s    
   </td>
   <td>
   </td>
   <td>
   </td>
  </tr>
  <tr>
   <td>2
   </td>
   <td>469s
   </td>
   <td>487s
   </td>
   <td>
   </td>
  </tr>
  <tr>
   <td>4
   </td>
   <td>473s
   </td>
   <td>570s
   </td>
   <td>
   </td>
  </tr>
  <tr>
   <td>8
   </td>
   <td>505s
   </td>
   <td>
   </td>
   <td>
   </td>
  </tr>
</table>

This is a single thread run of ~120 shaders.  

For the multithreaded run the results were the similar.  The time went from ~62sec to ~38sec.  The
optimal configuration was an 2 for the OperandData and an
std::vector for the OperandList.  Similar
expiriments were done with other sets of shaders.  The compile time still
improved, but not as much.

Contributes to https://github.com/KhronosGroup/SPIRV-Tools/issues/1609.