@dj2 2018-06-07
Out of curiosity, did you try calling .reserve() on the vector to set the number of operands it would need to store or are they just added to the back?