@ratchetfreak 2018-06-08
The unique_ptr to a std::vector seems like a waste (double indirection), checking against empty is just as fast as checking a pointer so you may as well keep it inline if you don't mind the extra space needed. Also std::aligned_storage is a better solution than a char array.

You are likely to get better performance for indexing and push_back by having a `iterator begin;` as a field and use that for the operator[]. You can even use that field to check whether it is pointing to the small buffer as the test to see whether the large or small buffer is being used.

This lets you recycle the storage space:

```
T* begin;//always points to &sb or bb.data()
size_t length;
union {
    std::aligned_storage<sizeof(T)*small_size, alignof(T)> sb;
    std::vector<T> bb;
}
```

This does require more lifetime finagling but moveToBig is then a std::move of sb to a local std::vector, destruct of the sb elements and then a placement move construct of local to bb.