@s-perron 2018-06-11
@ratchetfreak I implemented you suggestion (or a least part of it) in https://github.com/s-perron/SPIRV-Tools/tree/small_vector2.  It did not give any improvments.  In general I agree it would be better to reuse the same data; however, in this case I do not believe the amount of data being allocated is the bottleneck.  Even large shaders are not that large.  We just want to avoid a lot of calls to malloc and free.

For now, I will go with the simpler implementation.  I do not like unions all that much because the make it harder for the compiler to optimize, and easier for the developer to make a mistake.

I also tried Abseil, and it does give a small improvement over my implementation, but it is not very much.  So there is room for improvement.  We can also look into this a little more later, but I want to get the majority of the improvement first, then we can work to get the rest.