@dneto0 2018-06-11
Overall the approach seems fine and the code looks good.

However, bots fail compilation:

/home/travis/build/KhronosGroup/SPIRV-Tools/source/util/small_vector.h:336:12: error: variable has incomplete type 'void'
      auto ret = large_data_->insert(new_pos, first, last);