@dneto0 2018-06-11
At one point I wondered if the creation of operand lists is sufficiently constrained that in the vast majority of cases we have all the operand values in hand at one time, so we can allocate their storage in bulk.  But that's a more intrusive change to this one.